refactor: Item について、別クラスからも参照されておりファイルを分けた方が可読性が高くなりそうだったため分けた。
